Allow passing --local flag to booster/merge.

+++ booster/merge.zsh
@@ -13,12 +13,14 @@ function booster_merge {
       echo -e $red"There are merge conflicts with master"
     else
       git fetch
-      git push --force-with-lease --quiet origin $current_branch
+      if [ "$1" != "--local" ]; then
+        git push --force-with-lease --quiet origin $current_branch
+      fi
 
       number_of_commits=$(git log HEAD...origin/master --pretty=oneline | wc -l)
       if [ "$number_of_commits" -gt 1 ]; then
         echo $red"You have multiple commit messages. It's advisable to squash all commits into just one."
-        echo $red"\n$(git log HEAD...origin/master --pretty=oneline | cut -d' ' -f2-)\n"
+        echo $red"\n$(git log HEAD...origin/master --pretty=oneline --reverse | cut -d' ' -f2-)\n"
         echo "Do you still want to continue merging? (Y/N)"
         read mergeable
       fi
@@ -34,11 +36,15 @@ function booster_merge {
         git push origin master
 
         echo -e $green"Cleaning up"
-        git push origin --quiet :$current_branch # delete remote branch
+        if [ "$1" != "--local" ]; then
+          git push origin --quiet :$current_branch # delete remote branch
+        fi
         git branch -D --quiet $current_branch # delete local branch
 
         echo -e $green"\n$current_branch has been merged to master"
-        echo -e $green"$current_branch has been deleted from remote repository"
+        if [ "$1" != "--local" ]; then
+          echo -e $green"$current_branch has been deleted from remote repository"
+        fi
         echo -e $green"$current_branch has been deleted from local repository\n"
       else
         echo $red"Merge aborted..."
